refactor: use hybrid architecture to avoid CORS issue | SCRUM-499
